refactored all IDs from int to Long. Did all the way to Lab 7 (included)

+++ Lab-07-solution/src/main/java/com/petclinic/core/OwnerService.java
@@ -20,7 +20,7 @@ public class OwnerService {
         else return owner.get();
     }
 
-    public Owner findById(Integer id) {
+    public Owner findById(Long id) {
         var owner = ownerRepository.findById(id);
         if (owner.isEmpty()) {
             throw new EntityNotFoundException("could not find Owner with id: " + id);
